requirements.txt
fastapi==0.128.5
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic[email]
PyJWT==2.10.1
python-dotenv==1.0.1
mysql-connector-python==9.2.0
cryptography==46.0.4
email-validator==2.3.0
slowapi==0.1.8
requests==2.32.3
aiohttp==3.13.3
livekit==1.0.25
livekit-agents==1.4.1
livekit-api==1.1.0
livekit-protocol==1.1.2